v0.0.5: - Adjusted generation of all structures to be more common - Reengineered nether fortress jigsaw generation yet again - Renamed wither forts to catacombs - Adjusted piglin manor chest generation - Added soul stone blobs to soul sand valleys - Added spawn eggs for piglin hunters - Added integration with Biomes O'Plenty - Added integration with Amplified Nether - Added entity and item names in lang - Many code refinements

+++ src/main/java/com/izofar/bygonenether/entity/PiglinPrisoner.java
@@ -82,7 +82,7 @@ public class PiglinPrisoner extends AbstractPiglin implements CrossbowAttackMob,
 	public static AttributeSupplier.Builder createAttributes() { 
 		return Monster.createMonsterAttributes()
 				.add(Attributes.MAX_HEALTH, 24.0D)
-				.add(Attributes.MOVEMENT_SPEED, (double) 0.35F)
+				.add(Attributes.MOVEMENT_SPEED, 0.35F)
 				.add(Attributes.ATTACK_DAMAGE, 6.0D); 
 		}
